Work in progress, implmenting game logic in the reducers and adding a board display in the CellGrid components
Its challenging getting good typesafely around the redux/react-redux
boundries. i.e. getting the actions typed in the reducers still feels messy.
Also the connect() method when connecting a pure display component into
redux feels error prone. I split the props interface out to allow some
type safety, but this was very manual and it
was happy with some implicit any's and errors.
Also Im not happy that the arguments to the display component are
impacted by redux. To me it feels like the CellProps should be owned by
the Cell.tsx, and the container should just have to match these by
combining mapState and mapDIspatch, right now the standard seems to be let
react-redux model blead into the props of the non react-redux component